+++ Android/app/src/main/java/com/example/rocketcorner/fragments/ShopFragment.java
@@ -1,6 +1,5 @@
 package com.example.rocketcorner.fragments;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.annotation.Nullable;
@@ -13,9 +12,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.example.rocketcorner.R;
-import com.example.rocketcorner.StoreActivity;
 import com.example.rocketcorner.adapters.ItemAdapter;
-import com.example.rocketcorner.profileActivity;
 
 import java.util.ArrayList;
 import java.util.List;
