+++ Android/app/src/main/res/layout/activity_main.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:background="@drawable/rocketc"
-    tools:context=".MainActivity"
+    tools:context=".LoginActivity"
     android:layout_height="match_parent"
     android:layout_width="match_parent">
