+++ Android/app/src/main/res/layout/activity_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".profileActivity">
+    tools:context=".MainActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
